refactor(web): step 7 — extract pending-ACP session staging to store/chat/acp-staging.ts
